// src/api/generated/model/pagedListMgmtSoftwareModuleType.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * hawkBit REST APIs
 * Eclipse hawkBit™ is a domain-independent back-end framework for rolling out software updates to constrained edge devices as well as more powerful controllers and gateways connected to IP based networking infrastructure.

 * OpenAPI spec version: v1
 */
import type { MgmtSoftwareModuleType } from './mgmtSoftwareModuleType';
import type { Links } from './links';

export interface PagedListMgmtSoftwareModuleType {
  content: MgmtSoftwareModuleType[];
  total?: number;
  size?: number;
  _links?: Links;
}