// src/api/generated/model/exceptionInfo.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * hawkBit REST APIs
 * Eclipse hawkBit™ is a domain-independent back-end framework for rolling out software updates to constrained edge devices as well as more powerful controllers and gateways connected to IP based networking infrastructure.

 * OpenAPI spec version: v1
 */
import type { ExceptionInfoInfo } from './exceptionInfoInfo';

export interface ExceptionInfo {
  exceptionClass?: string;
  errorCode?: string;
  message?: string;
  info?: ExceptionInfoInfo;
}